Use shader in SurfaceVis
