+++ packages/lib/src/vis/heatmap/hooks.ts
@@ -5,9 +5,11 @@ import {
   getSafeDomain,
   getPixelEdgeValues,
   toTextureSafeNdArray,
+  getDataTexture,
 } from './utils';
 
 export const useVisDomain = createMemo(getVisDomain);
 export const useSafeDomain = createMemo(getSafeDomain);
 export const usePixelEdgeValues = createMemo(getPixelEdgeValues);
 export const useTextureSafeNdArray = createMemo(toTextureSafeNdArray);
+export const useDataTexture = createMemo(getDataTexture);
